+++ plugin_ide.ui/src-lang/melnorme/lang/ide/ui/editor/actions/AbstractEditorHandler.java
@@ -13,21 +13,32 @@ package melnorme.lang.ide.ui.editor.actions;
 
 import static melnorme.utilbox.core.Assert.AssertNamespace.assertNotNull;
 import melnorme.lang.ide.ui.EditorSettings_Actual;
-import melnorme.lang.ide.ui.LangUIMessages;
+import melnorme.lang.ide.ui.editor.AbstractLangEditor;
 
 import org.eclipse.core.commands.AbstractHandler;
 import org.eclipse.core.commands.ExecutionEvent;
 import org.eclipse.core.commands.ExecutionException;
+import org.eclipse.swt.widgets.Shell;
 import org.eclipse.ui.IWorkbenchPage;
+import org.eclipse.ui.editors.text.TextEditorActionContributor;
 import org.eclipse.ui.handlers.HandlerUtil;
 import org.eclipse.ui.texteditor.ITextEditor;
 
+/**
+ * This is a an editor handler that manages its enablement state so as to be restricted to a certaint type
+ * of editors only. It is meant to be used by {@link TextEditorActionContributor}'s, for which there is 
+ * a single instance per multiple editor ids. 
+ */
 public abstract class AbstractEditorHandler extends AbstractHandler {
 	
 	protected final IWorkbenchPage page;
 	
 	public AbstractEditorHandler(IWorkbenchPage page) {
 		this.page = assertNotNull(page);
+	}
+	
+	protected Shell getShell() {
+		return page.getWorkbenchWindow().getShell();
 	}
 	
 	@Override
@@ -38,14 +49,18 @@ public abstract class AbstractEditorHandler extends AbstractHandler {
 	@Override
 	public Object execute(ExecutionEvent event) throws ExecutionException {
 		ITextEditor editor = (ITextEditor) HandlerUtil.getActiveEditorChecked(event);
-		if(editor == null) {
-			throw new ExecutionException(LangUIMessages.INTERNAL_ERROR_NoEditorForCommand);
-		}
-		runWithEditor(editor);
+		
+		new LangEditorRunner(getShell(), editor) {
+			
+			@Override
+			protected void runWithLangEditor(AbstractLangEditor editor) {
+				doRunWithEditor(editor);
+			}
+		}.run();
 		
 		return null;
 	}
 	
-	public abstract void runWithEditor(ITextEditor editor);
+	protected abstract void doRunWithEditor(AbstractLangEditor editor);
 	
 }